[#420] style : userInfo 디자인 변경
