+++ components/common/userInfo/index.tsx
@@ -1,0 +1,3 @@
+import UserInfo from "./userInfo";
+
+export default UserInfo;
